Added versionning to the GTM busting file (#1474)
* Added versionning to the GTM busting file
* move common code to a trait
